Initial commit of the 'smallcheck' library
This is the first commit of this project. It is not functional.
There's a note left in `src/serial.rs` as to why. Getting the Serial
trait correct will be the tricky bit here, everything else is
friendly window-dressing, at least in so far as I am aware from this
place of wild-eyed naivety.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>